+++ Assignment6/RendererConsole.cpp
@@ -23,9 +23,9 @@ void RendererConsole::render(const LifeSimulator& simulation)
     {
         populate(simulation.getSizeX(), simulation.getSizeX());
     }
-    for (int i = 0; i < simulation.getSizeX(); i++)
+    for (uint8_t i = 0; i < simulation.getSizeX(); i++)
     {
-        for (int j = 0; j < simulation.getSizeY(); j++)
+        for (uint8_t j = 0; j < simulation.getSizeY(); j++)
         {
             if (simulation.getCell(i, j) != m_state[i][j])
             {
